Futures as queue and search attendance by date
Where last commit changed the way futures are queued -- this is reverted in this commit. It is similar to previous implementation and changes to some parts of the code fixed the memory bloat for large searches. Workers allocated to thread pool is now fixed at 64 -- possibly will change in the future. In this commit, the user will no longer need to do the guesswork of setting the range of timetable_id as an algorithm for searching the range of timetable_id by range of date is now added.