Added mitigation against unsorted timetable_id
Rarely, some timetable_id isn't generated at the same time as its peers for a day in date, maybe due to manual intervention. This causes the backlogged timetable_id to be generated later, usually after a timetable_id for the following date. Typically, timetable_ids are sorted by date for the whole range of timetable_id. This commit adds an option to induce dateToTimetableID() to call itself to search for timetable_ids beyond the local interval of timetable_id it found.