Sped up registered class lookup
Instead of parsing through lists of students in each classes, the script now probes class_id(s) with POST requests to https://mmls.mmu.edu.my/attendancelogin? to check if the student is registered in that class or not. Apparently, the attendance system checks whether the student ID is present in the class first before verifying the student's password which is somehow quicker than loading and parsing through student lists. Also, as the lookup time is halved, the user is no longer asked whether they want to let the script pre-fill registered classes for them. It does regardless.