Added warning and asks to autofill
Date search malfunctions for the first three days of trimester, or possibly in the first week, due to date-unsorted attendance links. They appear to be pregenerated three days ahead of time. I have added a warning in the date search option that it is inaccurate. Also, the script no longer defaults to autofill classes and instead asks the user if they want to. Additionally, time logging is changed.