More warnings and checking!
Script now logs into MMU Mobile and grabs the trimester start date which is now being checked against if the start date is out of range. If the start date is within the first three days of a trimester, the script displays a scaaary warning, warning the user to expect missing attendance links. Also, I changed some wordings.