Now using asyncio and aiohttp for concurrency
The script used to use concurrent.futures and schedule futures in threads. This approach has a significant overhead which causes CPU usage to balloon as more connections are scheduled at a time. This is exacerbated by the GIL causing threads to effectively run single-threaded where each thread spend most of the time waiting for lock to be acquired, and the standard library's urllib which does not support the keep-alive header, leading to unnecessarily tearing down and rebuilding connection for each request. Using asyncio significantly reduces CPU load, and because urllib is incompatible, the use of aiohttp in its stead -- which supports keep-alive -- increases throughput even with less concurrent requests scheduled at a time. aiohttp is now a dependency as it is not in standard library.