Considerable changes to mmlsattendance module
I don't remember all changes off the top of my head, but what I do remember: A list of tasks is no longer created when scraping. Instead, a feeder tasks is created which enqueues tasks to a queue of size 500. This change is so that memory usage don't bloat for scraping over a wide range. Caching is also implemented and used if the CONNECT_DB flag in the module is set to True. With it enabled, during the fetching of attendance URLs over a range of timetable ID, if URL returns 200 OK, it caches the input fields of the attendance sign in form to an SQLite3 database. The next time searching is done, it grabs the required information (form input fields) in the attendance URL from the database. If caching need to be executed ahead of time, a caching function (coroutine) is implemented. Before fetching from cache is done, scraping function (coroutine) will check if existing attendance URLs online was cleared, thus invalidating the cache. If it is, it removes everything in cache. The module may return MMLSResponseError when MMLS is having problems.