Replaced original discord bot
A rewrite of the original bot. Now, attendance scraping related commands are separated in its own cog, and any interaction with the mmlsattendance module is in the mmlscog extension. By default, the bot starts a background caching task that is scheduled to run every two hours, and if certain cases are true (i.e. after 12PM MST), it then runs every 6 hours which is then reset to 2 hours after midnight in MST. Commands that are meant to be run sequentially is blocked when other command of its like is running. Additionally, the user can cancel their running command if so desired. There are commands that are limited to the owner of the bot.