Significant rewrite to mmlsattendance.py
MMLS-Attendance-Scraper had been improved off repo and wasn't updated in a long while. Throughout the time it had gained caching capabilities using Sqlite3 and an option to serialize SubjectDB into a json-formatted string to facilitate saving. I have written below with the assumption that the repo was close enough to the current version. Only after I have finished written it that I realized I was wrong.

Changes:
The module has been significantly rewritten. The module no longer creates its own sqlite3 DB for caching, but rather stores the cache in-memory, with the option to load or return json-formatted str of the cache.

On the other hand, scrape_date() has an option to search a different way. Instead of binary searching exact timetable IDs a date range begins and ends, it binary searches any timetable ID that is in the range of 1 to 2 months prior to the specified start date. From that point, it linear searches forward. This is obviously slower, but looks to be way more reliable and faster than linear searching from MIN_TIMETABLE_ID.

Both scrape_date() and scrape() has now become an async generator instead of a coroutine. This is done because yielding found attendances as it is found is more efficient than storing them in a list before returning the whole lot to the user. Before this an asyncio.Queue() is passed so that results can be obtained as the coroutine is working. It is now deprecated.

Scraping functions/methods were moved into the Scraper class.

SubjectDB, InputFields and ScrapedTimetable were renamed into Courses, AttendanceForm and DetailedAttendanceForm respectively.

Type hints are now used throughout the code. It should be now easier to come back to and understand.